feat: split events index into heartbeat-events and core-events
This change introduces a new indexing strategy for incoming events to improve long-term data
retention and query performance.

- Events of type `HEARTBEAT` are now indexed under `heartbeat-events`.
- All other event types are indexed under `core-events`.
- A new ISM policy has been provisioned for each index:
  - `heartbeat-events` are subject to a 7-day TTL, aligned with the previous behavior.
  - `core-events` are retained for up to 2 years.
- The `user_events` alias now points to both `heartbeat-events` and `core-events`.
- A new alias `core_events` has been created to simplify queries targeting only core event types.
  This is intended to reduce overhead for queries that don't need heartbeat data.

BREAKING CHANGE: The unified "events" index is no longer used. Integrations depending on the
previous index or data structure must be updated.

+++ src/main/kotlin/ch/srgssr/pillarbox/monitoring/event/repository/OpenSearchConfigurationProperties.kt
@@ -19,7 +19,7 @@ data class OpenSearchConfigurationProperties(
   val uri: URI = URI("http://localhost:9200"),
   @NestedConfigurationProperty
   val retry: RetryProperties = RetryProperties(),
-  val timeout: Long = 10_000,
+  val timeout: Int = 10_000,
 ) {
   /**
    * Retrieves the host and port in the format `host:port` based on the URI.
